Move multiplicy column code to utils
