+++ stwo_cairo_prover/crates/prover/src/components/verify_instruction/prover.rs
@@ -32,8 +32,9 @@ use stwo_prover::core::utils::bit_reverse_coset_to_circle_domain_order;
 use stwo_prover::core::vcs::blake2_merkle::{Blake2sMerkleChannel, Blake2sMerkleHasher};
 
 use super::component::{Claim, InteractionClaim};
+use crate::components::utils::pack_values;
 use crate::components::{
-    memory_address_to_id, memory_id_to_big, pack_values, range_check_4_3, range_check_7_2_5,
+    memory_address_to_id, memory_id_to_big, range_check_4_3, range_check_7_2_5,
 };
 use crate::relations;
